Clippy/fmt

+++ rmp/src/encode/uint.rs
@@ -1,5 +1,5 @@
-use super::{RmpWrite, write_marker};
-use crate::encode::{ValueWriteError};
+use super::{write_marker, RmpWrite};
+use crate::encode::ValueWriteError;
 use crate::Marker;
 
 /// Encodes and attempts to write an unsigned small integer value as a positive fixint into the
@@ -127,11 +127,11 @@ pub fn write_u64<W: RmpWrite>(wr: &mut W, val: u64) -> Result<(), ValueWriteErro
 /// See [`write_uint`] for more info.
 pub fn write_uint8<W: RmpWrite>(wr: &mut W, val: u8) -> Result<Marker, ValueWriteError<W::Error>> {
     if val < 128 {
-        write_pfix(wr, val as u8)
-            .and(Ok(Marker::FixPos(val as u8)))
+        write_pfix(wr, val)
+            .and(Ok(Marker::FixPos(val)))
             .map_err(ValueWriteError::InvalidMarkerWrite)
     } else {
-        write_u8(wr, val as u8).and(Ok(Marker::U8))
+        write_u8(wr, val).and(Ok(Marker::U8))
     }
 }
